% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salOrder.R
\name{salOrderUploadServer}
\alias{salOrderUploadServer}
\title{处理逻辑（带进度条）}
\usage{
salOrderUploadServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
处理逻辑（带进度条）
}
\examples{
salOrderUploadServer()
}
